### 2023-09-02 #5797 @editorialbot
[PRE REVIEW]: kinisi: Bayesian analysis of mass transport from molecular dynamics simulations
**Submitting author:** @arm61 (<a href="http://orcid.org/0000-0003-3381-5911">Andrew McCluskey</a>)
**Repository:** https://github.com/bjmorgan/kinisi/
**Branch with paper.md** (empty if default branch):
**Version:** 0.7.0
