### 2023-09-02 #5797 @arm61
Also, just to note. The title of the paper has changed to the more descriptive “kinisi: Bayesian analysis of mass transport from molecular dynamics simulations”.
