### 2023-10-17 f1c169e
First version of posterior predictive implementation
