### 2023-10-23 c55ff87
Use uravu 1.3.0 to sample the posterior predictive of the arrhenius
