+++ pyproject.toml
@@ -35,7 +35,7 @@ dependencies = [
     'scikit-learn',
     'scipy>=1.9.3',
     'tqdm',
-    'uravu>=1.2.9'
+    'uravu>=1.3.0'
 ]
 
 [project.optional-dependencies]
