+++ kinisi/arrhenius.py
@@ -155,7 +155,7 @@ class SuperArrhenius(Relationship):
         """
         return self.function(self.x[:, np.newaxis], self.flatchain[:, 0], self.flatchain[:, 1], self.flatchain[:, 2])
 
-    def extrapolate(self, extrapolated_temperature: float) -> 'uravu.distribution.Distribution':
+    def extrapolate(self, extrapolated_temperature: Union[float, List[float], np.ndarray], posterior_predictive_kwargs=None) -> 'uravu.distribution.Distribution':
         """
         Extrapolate the diffusion coefficient to some un-investigated value. This can also be
         used for interpolation.
@@ -163,7 +163,13 @@ class SuperArrhenius(Relationship):
         :param: Temperature to return diffusion coefficient at.
         :return: Diffusion coefficient at extrapolated temperature.
         """
-        return self.function(extrapolated_temperature, self.flatchain[:, 0], self.flatchain[:, 1], self.flatchain[:, 2])
+        if posterior_predictive_kwargs is None:
+            posterior_predictive_kwargs = {}
+        if isinstance(extrapolated_temperature, (float, int, complex)) :
+            extrapolated_temperature = np.array([extrapolated_temperature])
+        if isinstance(extrapolated_temperature, List):
+            extrapolated_temperature = np.array(extrapolated_temperature)
+        return self.posterior_predictive(abscissa_values=extrapolated_temperature, **posterior_predictive_kwargs)
 
 
 def super_arrhenius(abscissa: np.ndarray, activation_energy: float, prefactor: float, t_zero: float) -> np.ndarray:
