### 2023-10-23 7bad5f0
Int now valid for get_indices
Added parsing for interger indexes to the get_indices function of the MDAnalysisParser class
