+++ kinisi/parser.py
@@ -207,6 +207,7 @@ class Parser:
             n_samples = np.append(n_samples, disp.shape[0] * timesteps[-1] / timestep)
         return delta_t, disp_3d, n_samples
 
+
 class ASEParser(Parser):
     """
     A parser for ASE Atoms objects
@@ -236,7 +237,8 @@ class ASEParser(Parser):
         gigabytes (GB). Optional, defaults to :py:attr:`8.`.
     :param progress: Print progress bars to screen. Optional, defaults to :py:attr:`True`.
     """
-    def __init__(self, 
+
+    def __init__(self,
                  atoms: List["ase.atoms.Atoms"],
                  specie: str,
                  time_step: float,
@@ -249,31 +251,20 @@ class ASEParser(Parser):
                  sampling: str = 'multi-origin',
                  memory_limit: float = 8.,
                  progress: bool = True):
-        
-            structure, coords, latt = self.get_structure_coords_latt(atoms, sub_sample_traj, progress)
-            indices = self.get_indices(structure, specie)
 
-            super().__init__(self.get_disp(coords, latt),
-                         indices[0], 
-                         indices[1], 
-                         time_step, 
-                         step_skip, 
-                         min_dt, 
-                         max_dt, 
-                         n_steps, 
-                         spacing, 
-                         sampling, 
-                         memory_limit, 
-                         progress)
-            self._volume = structure.get_volume()
-            
-        
+        structure, coords, latt = self.get_structure_coords_latt(atoms, sub_sample_traj, progress)
+        indices = self.get_indices(structure, specie)
+
+        super().__init__(self.get_disp(coords, latt), indices[0], indices[1], time_step, step_skip, min_dt, max_dt,
+                         n_steps, spacing, sampling, memory_limit, progress)
+        self._volume = structure.get_volume()
+
     @staticmethod
     def get_structure_coords_latt(
-                atoms: List["ase.atoms.Atoms"],
-                sub_sample_traj: int = 1,
-                progress: bool = True) -> Tuple["ase.atoms.Atoms", List[np.ndarray], List[np.ndarray]]:
-            """
+            atoms: List["ase.atoms.Atoms"],
+            sub_sample_traj: int = 1,
+            progress: bool = True) -> Tuple["ase.atoms.Atoms", List[np.ndarray], List[np.ndarray]]:
+        """
             Obtain the initial structure and displacement from a :py:attr:`list`
             of :py:class`pymatgen.core.structure.Structure`.
 
@@ -285,29 +276,26 @@ class ASEParser(Parser):
             :return: Tuple containing: initial structure, fractional coordinates for all atoms,
                 and lattice descriptions.
             """
-            coords, latt = [], []
-            first = True
-            if progress:
-                iterator = tqdm(atoms[::sub_sample_traj], desc='Reading Trajectory')
-            else:
-                iterator = atoms[::sub_sample_traj]
-            for struct in iterator:
-                if first:
-                    structure = struct
-                    first = False
-                scaled_positions = struct.get_scaled_positions()
-                coords.append(np.array(scaled_positions)[:, None])
-                latt.append(struct.cell[:])
-            
-            coords.insert(0, coords[0])
-            latt.insert(0, latt[0])
-            return structure, coords, latt
-    
+        coords, latt = [], []
+        first = True
+        if progress:
+            iterator = tqdm(atoms[::sub_sample_traj], desc='Reading Trajectory')
+        else:
+            iterator = atoms[::sub_sample_traj]
+        for struct in iterator:
+            if first:
+                structure = struct
+                first = False
+            scaled_positions = struct.get_scaled_positions()
[…]
