### 2023-10-24 #5797 @zhubonan
Hi @dengzeyu, @shyuep would you be interested to review this JOSS submission? It is about using Bayesian analysis to obtain mass-transports coefficients and their error bars from MD simulation trajectories, which I think are relevant to your areas of expertise.

The manuscript can be viewed here 👉 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.05797/joss.05797/10.21105.joss.05797.pdf)
Software repository 👉 https://github.com/bjmorgan/kinisi/
The JOSS reviewer policy and reviewer guideline for your reference 👉 [Reviewer Guideline](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Thanks!
