### 2023-10-24 99c56c0
New Keyword: s_indices added
Added a new keyword to MDAnalysisParser: s_indices
