### 2023-10-24 70ca0f7
Extended indices to other parsers
Extended the s_indices keyword to ASEParser and PymatgenParser
