+++ kinisi/tests/test_parser.py
@@ -98,6 +98,22 @@ class TestParser(unittest.TestCase):
         assert_almost_equal(data.step_skip, 50)
         assert_equal(data.indices, list(range(xd.natoms[0])))
 
+    def test_pymatgen_init_with_indices(self):
+        xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
+        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 's_indices': [3, 4, 5]}
+        data = parser.PymatgenParser(xd.structures, **da_params)
+        assert_almost_equal(data.time_step, 2.0)
+        assert_almost_equal(data.step_skip, 50)
+        assert_equal(data.indices, [3, 4, 5])
+
+    def test_pymatgen_init_with_molecules(self):
+        xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
+        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 's_indices': [[3, 4, 5], [6, 7]]}
+        data = parser.PymatgenParser(xd.structures, **da_params)
+        assert_almost_equal(data.time_step, 2.0)
+        assert_almost_equal(data.step_skip, 50)
+        assert_equal(data.indices, [3, 4, 5, 6, 7])
+
     def test_pymatgen_big_timestep(self):
         xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
         da_params = {'specie': 'Li', 'time_step': 20.0, 'step_skip': 100}
@@ -113,6 +129,22 @@ class TestParser(unittest.TestCase):
         assert_almost_equal(data.time_step, 1e-3)
         assert_almost_equal(data.step_skip, 1)
         assert_equal(data.indices, list(range(180)))
+
+    def test_ase_init_with_indices(self):
+        traj = Trajectory(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_ase.traj'))
+        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 's_indices': [100, 101, 90]}
+        data = parser.ASEParser(traj, **da_params)
+        assert_almost_equal(data.time_step, 1e-3)
+        assert_almost_equal(data.step_skip, 1)
+        assert_equal(data.indices, [100, 101, 90])
+
+    def test_ase_init_with_molecules(self):
+        traj = Trajectory(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_ase.traj'))
+        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 's_indices': [[100, 101], [103]]}
+        data = parser.ASEParser(traj, **da_params)
+        assert_almost_equal(data.time_step, 1e-3)
+        assert_almost_equal(data.step_skip, 1)
+        assert_equal(data.indices, [100, 101, 103])
 
     def test_mda_init(self):
         xd = mda.Universe(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_LAMMPS.data'),
