### 2023-10-25 #5984 @editorialbot
[REVIEW]: kinisi: Bayesian analysis of mass transport from molecular dynamics simulations 
**Submitting author:** @arm61 (<a href="http://orcid.org/0000-0003-3381-5911">Andrew McCluskey</a>)
**Repository:** https://github.com/bjmorgan/kinisi/
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
