### 2023-10-25 #5984 @zhubonan
Hi @denzeyu & @hmacdope, thanks again for agreeing to review this package.

First of all, please run this command to generate the review checklist:

and tick the boxes as you go through the submission.

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns, please let me know.

We are aiming to have the review done in two weeks if possible - If you need more time, please let me know here in this issue.

For comments and queries, you can open issues/PRs directly in the package's repository. Please mention this issue in the tickets opened in order to keep tracking.
