### 2023-10-26 #5984 @zhubonan
My guess is that the metadata is stored based on the title upon submission. It should affect the review process. We just to the make sure it is corrected in the system before acceptance.
