### 2023-11-01 44a38c5
Added docstrings
