+++ kinisi/tests/test_parser.py
@@ -100,7 +100,7 @@ class TestParser(unittest.TestCase):
 
     def test_pymatgen_init_with_indices(self):
         xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
-        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 's_indices': [3, 4, 5]}
+        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 'specie_indices': [3, 4, 5]}
         data = parser.PymatgenParser(xd.structures, **da_params)
         assert_almost_equal(data.time_step, 2.0)
         assert_almost_equal(data.step_skip, 50)
@@ -108,11 +108,12 @@ class TestParser(unittest.TestCase):
 
     def test_pymatgen_init_with_molecules(self):
         xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
-        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 's_indices': [[3, 4, 5], [6, 7]]}
+        molecules = [[2, 3, 4], [5, 6, 7]]
+        da_params = {'specie': None, 'time_step': 2.0, 'step_skip': 50, 'specie_indices': molecules}
         data = parser.PymatgenParser(xd.structures, **da_params)
         assert_almost_equal(data.time_step, 2.0)
         assert_almost_equal(data.step_skip, 50)
-        assert_equal(data.indices, [3, 4, 5, 6, 7])
+        assert_equal(data.indices, list(range(len(molecules))))
 
     def test_pymatgen_big_timestep(self):
         xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
@@ -132,7 +133,7 @@ class TestParser(unittest.TestCase):
 
     def test_ase_init_with_indices(self):
         traj = Trajectory(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_ase.traj'))
-        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 's_indices': [100, 101, 90]}
+        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 'specie_indices': [100, 101, 90]}
         data = parser.ASEParser(traj, **da_params)
         assert_almost_equal(data.time_step, 1e-3)
         assert_almost_equal(data.step_skip, 1)
@@ -140,11 +141,12 @@ class TestParser(unittest.TestCase):
 
     def test_ase_init_with_molecules(self):
         traj = Trajectory(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_ase.traj'))
-        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 's_indices': [[100, 101], [103]]}
+        molecules = [[2, 3, 4], [5, 6, 7]]
+        da_params = {'specie': None, 'time_step': 1e-3, 'step_skip': 1, 'specie_indices': molecules}
         data = parser.ASEParser(traj, **da_params)
         assert_almost_equal(data.time_step, 1e-3)
         assert_almost_equal(data.step_skip, 1)
-        assert_equal(data.indices, [100, 101, 103])
+        assert_equal(data.indices, list(range(len(molecules))))
 
     def test_mda_init(self):
         xd = mda.Universe(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_LAMMPS.data'),
@@ -160,7 +162,7 @@ class TestParser(unittest.TestCase):
[…]
