+++ kinisi/parser.py
@@ -251,19 +251,25 @@ class ASEParser(Parser):
                  sampling: str = 'multi-origin',
                  memory_limit: float = 8.,
                  progress: bool = True,
-                 s_indices: List[int] = None):
+                 specie_indices: List[int] = None,
+                 center: str = 'Geometry',
+                 masses: List[float] = None):
 
         structure, coords, latt = self.get_structure_coords_latt(atoms, sub_sample_traj, progress)
 
         if specie != None:
             indices = self.get_indices(structure, specie)
-        elif isinstance(s_indices, list):
-            if isinstance(s_indices[0], list):
-                structure, coords, indices = self.get_molecules(structure, coords, s_indices)
+        elif isinstance(specie_indices, (list, tuple)):
+            if isinstance(specie_indices[0], (list, tuple)):
+                coords, indices = self.get_molecules(
+                    structure, coords, specie_indices, center,
+                    masses)  #Warning: This function changes the structure without changing the MDAnalysis object
             else:
-                indices = self.get_framework(structure, s_indices)
+                indices = self.get_framework(structure, specie_indices)
         else:
             raise TypeError('Unrecognized type for Specie or Indices')
+
+        self.coords_check = coords[0]
 
         super().__init__(self.get_disp(coords, latt), indices[0], indices[1], time_step, step_skip, min_dt, max_dt,
                          n_steps, spacing, sampling, memory_limit, progress)
@@ -329,23 +335,74 @@ class ASEParser(Parser):
         return indices, framework_indices
 
     @staticmethod
-    def get_molecules(structure: "ase.atoms.Atoms", coords: List[np.ndarray],
-                      indices: List[int]) -> Tuple[np.ndarray, np.ndarray, Tuple[np.ndarray, np.ndarray]]:
+    def get_molecules(structure: "MDAnalysis.universe.Universe", coords: List[np.ndarray], indices: List[int],
+                      center: str, masses: List[float]) -> Tuple[np.ndarray, np.ndarray, Tuple[np.ndarray, np.ndarray]]:
         """
+        Determine framework and non-framework indices for an :py:mod:`MDAnalysis` compatible file when 
+            specie_indices are provided and contain multiple molecules
         
+        Warning: This function changes the structure without changing the MDAnalysis object
+         
+        :param structure: Initial structure.
+        :param coords: fractional coordinates for all atoms.
+        :param indices: indices for the atoms in the molecules in the trajectory used in the calculation 
+            of the diffusion.
+        :param center: String describing center to calculate: Geometry or Mass.
+        :param masses: Masses associated with indices in indices.
+        
+        
+        :return: Tuple containing: Tuple containing: fractional coordinates for centers and framework atoms
+            and Tuple containing: indices for centers used in the calculation 
+            of the diffusion and indices of framework atoms.
         """
         framework_indices = []
-        flat_indices = np.concatenate(indices)
+        try:
+            indices = np.array(indices) - 1
+        except:
+            raise ValueError('Molecules must be of same length')
+
+        n_molecules = indices.shape[0]
+
         for i, site in enumerate(structure):
             if i not in indices:
                 framework_indices.append(i)
 
-        return structure, coords, (flat_indices, framework_indices)
+        if center == 'Geometry':
+            weights = None
+        elif center == 'Mass':
+            masses = np.array(masses)
+            weights = masses
+
+        sq_coords = np.squeeze(coords, axis=2)
+        s_coords = sq_coords[:, indices]
+        theta = s_coords * (2 * np.pi)
+        xi = np.cos(theta)
+        zeta = np.sin(theta)
+        xi_bar = np.average(xi, axis=-2, weights=weights)
[…]
