### 2023-11-05 28be826
Add docs installation
