### 2023-11-05 12bbb16
Restore methodology
