+++ kinisi/diffusion_analyzer.py
@@ -27,7 +27,7 @@ class DiffusionAnalyzer(Analyzer):
         There is one array in the list for each delta_t value. Note: it is necessary to use a list of arrays as
         the number of observations is not necessary the same at each data point.
     :param volume: The volume of the simulation cell.
-    :param bootstrap_params: The parameters for the :py:class:`kinisi.diffusion.DiffBootstrap` object. See
+    :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.DiffBootstrap` object. See
         the appropriate documentation for more guidance on this. Optional, default is the default bootstrap parameters.
     """
 
@@ -62,7 +62,7 @@ class DiffusionAnalyzer(Analyzer):
                                              List['pymatgen.core.structure.Structure']]],
                       parser_params: dict,
                       dtype: str = None,
-                      bootstrap_params: dict = None):
+                      uncertainty_params: dict = None):
         """
         Create a :py:class:`DiffusionAnalyzer` object from a list or nested list of
         :py:class:`pymatgen.core.structure.Structure` objects.
@@ -76,23 +76,23 @@ class DiffusionAnalyzer(Analyzer):
             these constitute a series of :py:attr:`consecutive` trajectories or a series of :py:attr:`identical`
             starting points with different random seeds, in which case the `dtype` should be either
             :py:attr:`consecutive` or :py:attr:`identical`.
-        :param bootstrap_params: The parameters for the :py:class:`kinisi.diffusion.MSDBootstrap` object. See the
+        :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.MSDBootstrap` object. See the
             appropriate documentation for more guidance on this. Optional, default is the default bootstrap parameters.
 
         :return: Relevant :py:class:`DiffusionAnalyzer` object.
         """
-        if bootstrap_params is None:
-            bootstrap_params = {}
+        if uncertainty_params is None:
+            uncertainty_params = {}
         diff = super()._from_pymatgen(trajectory, parser_params, dtype=dtype)
-        diff._diff = diffusion.MSDBootstrap(diff._delta_t, diff._disp_3d, diff._n_o, **bootstrap_params)
+        diff._diff = diffusion.MSDBootstrap(diff._delta_t, diff._disp_3d, diff._n_o, **uncertainty_params)
         return diff
-    
+
     @classmethod
     def from_ase(cls,
-                    trajectory: List[Union['ase.atoms.Atoms', List['ase.atoms.Atoms']]],
-                    parser_params: dict,
-                    dtype: str = None,
-                    bootstrap_params: dict = None):
+                 trajectory: List[Union['ase.atoms.Atoms', List['ase.atoms.Atoms']]],
+                 parser_params: dict,
+                 dtype: str = None,
+                 uncertainty_params: dict = None):
         """
         Create a :py:class:`DiffusionAnalyzer` object from a list or nested list of
         :py:class:`ase.atoms.Atoms` objects.
@@ -106,15 +106,15 @@ class DiffusionAnalyzer(Analyzer):
             these constitute a series of :py:attr:`consecutive` trajectories or a series of :py:attr:`identical`
             starting points with different random seeds, in which case the `dtype` should be either
             :py:attr:`consecutive` or :py:attr:`identical`.
-        :param bootstrap_params: The parameters for the :py:class:`kinisi.diffusion.MSDBootstrap` object. See the
+        :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.MSDBootstrap` object. See the
[…]
