### 2023-11-09 1ea0b19
Added error check for Molecules
