### 2023-11-10 04b4725
Added tests for COG and COM
