### 2023-11-14 39c37fc
Add catch for sub_sample_atoms and specie_indices
