### 2023-11-14 bf377a7
Added COG notebook
