+++ kinisi/parser.py
@@ -25,7 +25,7 @@ class Parser:
 
     :param disp: Displacements of atoms with the shape [site, time step, axis].
     :param indices: Indices for the atoms in the trajectory used in the diffusion calculation.
-    :param framework_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
+    :param drift_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
         calculation.
     :param time_step: Time step, in picoseconds, between steps in trajectory.
     :param step_skip: Sampling freqency of the trajectory (time_step is multiplied by this number to get the real
@@ -52,7 +52,7 @@ class Parser:
     def __init__(self,
                  disp: np.ndarray,
                  indices: np.ndarray,
-                 framework_indices: np.ndarray,
+                 drift_indices: np.ndarray,
                  time_step: float,
                  step_skip: int,
                  min_dt: float = None,
@@ -71,7 +71,7 @@ class Parser:
         self.sampling = sampling
         self._volume = None
 
-        drift_corrected = self.correct_drift(framework_indices, disp)
+        drift_corrected = self.correct_drift(drift_indices, disp)
         self.dc = drift_corrected
 
         if self.max_dt is None:
@@ -116,19 +116,19 @@ class Parser:
         return disp
 
     @staticmethod
-    def correct_drift(framework_indices: np.ndarray, disp: np.ndarray) -> np.ndarray:
+    def correct_drift(drift_indices: np.ndarray, disp: np.ndarray) -> np.ndarray:
         """
         Perform drift correction, such that the displacement is calculated normalised to any framework drift.
 
-        :param framework_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
+        :param drift_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
             calculation.
         :param disp: Numpy array of with shape [site, time step, axis] that describes the displacements.
 
         :return: Displacements corrected to account for drift of a framework.
         """
         # drift corrected position
-        if len(framework_indices) > 0:
-            framework_disp = disp[framework_indices]
+        if len(drift_indices) > 0:
+            framework_disp = disp[drift_indices]
             drift_corrected = disp - np.average(framework_disp, axis=0)[None, :, :]
         else:
             drift_corrected = disp
@@ -253,7 +253,8 @@ class ASEParser(Parser):
                  progress: bool = True,
                  specie_indices: List[int] = None,
                  center: str = 'Geometry',
-                 masses: List[float] = None):
+                 masses: List[float] = None,
+                 framework_indices: List[int] = None):
 
         structure, coords, latt = self.get_structure_coords_latt(atoms, sub_sample_traj, progress)
 
@@ -322,17 +323,17 @@ class ASEParser(Parser):
             and indices of framework atoms.
         """
         indices = []
-        framework_indices = []
+        drift_indices = []
         if not isinstance(specie, List):
             specie = [specie]
         for i, site in enumerate(structure):
             if site.symbol in specie:
                 indices.append(i)
             else:
-                framework_indices.append(i)
+                drift_indices.append(i)
         if len(indices) == 0:
             raise ValueError("There are no species selected to calculate the mean-squared displacement of.")
-        return indices, framework_indices
+        return indices, drift_indices
 
     @staticmethod
     def get_molecules(structure: "MDAnalysis.universe.Universe", coords: List[np.ndarray], indices: List[int],
@@ -352,7 +353,7 @@ class ASEParser(Parser):
             and Tuple containing: indices for centers used in the calculation 
             of the diffusion and indices of framework atoms.
[…]
