+++ kinisi/parser.py
@@ -62,6 +62,7 @@ class Parser:
                  sampling: str = 'multi-origin',
                  memory_limit: float = 8.,
                  progress: bool = True):
+
         self.time_step = time_step
         self.step_skip = step_skip
         self.indices = indices
@@ -252,7 +253,6 @@ class ASEParser(Parser):
                  memory_limit: float = 8.,
                  progress: bool = True,
                  specie_indices: List[int] = None,
-                 center: str = 'Geometry',
                  masses: List[float] = None,
                  framework_indices: List[int] = None):
 
@@ -263,7 +263,7 @@ class ASEParser(Parser):
         elif isinstance(specie_indices, (list, tuple)):
             if isinstance(specie_indices[0], (list, tuple)):
                 coords, indices = self.get_molecules(
-                    structure, coords, specie_indices, center,
+                    structure, coords, specie_indices,
                     masses)  #Warning: This function changes the structure without changing the MDAnalysis object
             else:
                 indices = self.get_framework(structure, specie_indices)
@@ -337,7 +337,7 @@ class ASEParser(Parser):
 
     @staticmethod
     def get_molecules(structure: "MDAnalysis.universe.Universe", coords: List[np.ndarray], indices: List[int],
-                      center: str, masses: List[float]) -> Tuple[np.ndarray, np.ndarray, Tuple[np.ndarray, np.ndarray]]:
+                      masses: List[float]) -> Tuple[np.ndarray, np.ndarray, Tuple[np.ndarray, np.ndarray]]:
         """
         Determine framework and non-framework indices for an :py:mod:`MDAnalysis` compatible file when specie_indices are provided and contain multiple molecules. Warning: This function changes the structure without changing the MDAnalysis object
          
@@ -345,7 +345,6 @@ class ASEParser(Parser):
         :param coords: fractional coordinates for all atoms.
         :param indices: indices for the atoms in the molecules in the trajectory used in the calculation 
             of the diffusion.
-        :param center: String describing center to calculate: Geometry or Mass.
         :param masses: Masses associated with indices in indices.
         
         
@@ -365,9 +364,11 @@ class ASEParser(Parser):
             if i not in indices:
                 drift_indices.append(i)
 
-        if center == 'Geometry':
+        if masses == None:
             weights = None
-        elif center == 'Mass':
+        elif len(masses) != indices.shape[-1]:
+            raise ValueError('Masses must be the same length as a molecule')
+        else:
             masses = np.array(masses)
             weights = masses
 
@@ -454,7 +455,6 @@ class PymatgenParser(Parser):
                  memory_limit: float = 8.,
                  progress: bool = True,
                  specie_indices: List[int] = None,
-                 center: str = 'Geometry',
                  masses: List[float] = None,
                  framework_indices: List[int] = None):
 
@@ -465,7 +465,7 @@ class PymatgenParser(Parser):
         elif isinstance(specie_indices, (list, tuple)):
             if isinstance(specie_indices[0], (list, tuple)):
                 coords, indices = self.get_molecules(
-                    structure, coords, specie_indices, center,
+                    structure, coords, specie_indices,
                     masses)  #Warning: This function changes the structure without changing the MDAnalysis object
             else:
                 indices = self.get_framework(structure, specie_indices)
@@ -554,7 +554,7 @@ class PymatgenParser(Parser):
 
     @staticmethod
     def get_molecules(structure: "MDAnalysis.universe.Universe", coords: List[np.ndarray], indices: List[int],
-                      center: str, masses: List[float]) -> Tuple[np.ndarray, np.ndarray, Tuple[np.ndarray, np.ndarray]]:
[…]
