### 2023-11-20 efd7239
Removed center keyword from tests
