### 2023-11-20 d8ffa1d
Added framework_indices keyword
By using the framework_indices keyword the framework can be defined explicitly.
