### 2023-11-21 fac7926
Added test for no drift correction
