### 2023-11-21 fea84e8
Revert "Merge branch 'COG' of https://github.com/jd15489/kinisi_indices into COG"
This reverts commit 899a201ab5243bbf27ba001d6e9726203a4d1e58 [formerly 5709d805d7696539b63334ef00581a23153d94ff], reversing
changes made to 1f5f711d368b001179a70c73c583c5518737bf00 [formerly fac79261737fc1442d4bdef22c3f7d1650191e28].


Former-commit-id: c409d7d0732e906e943fefd030ff6afac6849d9e
