+++ kinisi/parser.py
@@ -25,7 +25,7 @@ class Parser:
 
     :param disp: Displacements of atoms with the shape [site, time step, axis].
     :param indices: Indices for the atoms in the trajectory used in the diffusion calculation.
-    :param framework_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
+    :param drift_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
         calculation.
     :param time_step: Time step, in picoseconds, between steps in trajectory.
     :param step_skip: Sampling freqency of the trajectory (time_step is multiplied by this number to get the real
@@ -52,7 +52,7 @@ class Parser:
     def __init__(self,
                  disp: np.ndarray,
                  indices: np.ndarray,
-                 framework_indices: np.ndarray,
+                 drift_indices: np.ndarray,
                  time_step: float,
                  step_skip: int,
                  min_dt: float = None,
@@ -62,6 +62,7 @@ class Parser:
                  sampling: str = 'multi-origin',
                  memory_limit: float = 8.,
                  progress: bool = True):
+
         self.time_step = time_step
         self.step_skip = step_skip
         self.indices = indices
@@ -72,7 +73,7 @@ class Parser:
         self.sampling = sampling
         self._volume = None
 
-        drift_corrected = self.correct_drift(framework_indices, disp)
+        drift_corrected = self.correct_drift(drift_indices, disp)
         self.dc = drift_corrected
 
         if self.max_dt is None:
@@ -117,19 +118,19 @@ class Parser:
         return disp
 
     @staticmethod
-    def correct_drift(framework_indices: np.ndarray, disp: np.ndarray) -> np.ndarray:
+    def correct_drift(drift_indices: np.ndarray, disp: np.ndarray) -> np.ndarray:
         """
         Perform drift correction, such that the displacement is calculated normalised to any framework drift.
 
-        :param framework_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
+        :param drift_indices: Indices for the atoms in the trajectory that should not be used in the diffusion
             calculation.
         :param disp: Numpy array of with shape [site, time step, axis] that describes the displacements.
 
         :return: Displacements corrected to account for drift of a framework.
         """
         # drift corrected position
-        if len(framework_indices) > 0:
-            framework_disp = disp[framework_indices]
+        if len(drift_indices) > 0:
+            framework_disp = disp[drift_indices]
             drift_corrected = disp - np.average(framework_disp, axis=0)[None, :, :]
         else:
             drift_corrected = disp
@@ -237,6 +238,9 @@ class ASEParser(Parser):
     :param memory_limit: Upper limit in the amount of computer memory that the displacements can occupy in
         gigabytes (GB). Optional, defaults to :py:attr:`8.`.
     :param progress: Print progress bars to screen. Optional, defaults to :py:attr:`True`.
+    :param specie_indices: 
+    :param masses: 
+    :param framework_indices: 
     """
 
     def __init__(self,
@@ -253,8 +257,8 @@ class ASEParser(Parser):
                  memory_limit: float = 8.,
                  progress: bool = True,
                  specie_indices: List[int] = None,
-                 center: str = 'Geometry',
-                 masses: List[float] = None):
+                 masses: List[float] = None,
+                 framework_indices: List[int] = None):
 
         structure, coords, latt = self.get_structure_coords_latt(atoms, sub_sample_traj, progress)
 
@@ -263,12 +267,12 @@ class ASEParser(Parser):
         elif isinstance(specie_indices, (list, tuple)):
             if isinstance(specie_indices[0], (list, tuple)):
                 coords, indices = self.get_molecules(
-                    structure, coords, specie_indices, center,
+                    structure, coords, specie_indices,
[…]
