+++ kinisi/tests/test_parser.py
@@ -209,18 +209,13 @@ class TestParser(unittest.TestCase):
         assert_almost_equal(data.step_skip, 1)
         assert_equal(data.indices, [0])
         assert_almost_equal(data.coords_check, [[[0.382421597, 0.2087361, 0.2]]])
-        
+
     def test_mda_inti_with_empty_framwork_indices(self):
         xd = mda.Universe(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_LAMMPS_drift.data'),
                           os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_LAMMPS_drift.traj'),
                           topology_format='DATA',
                           format='LAMMPSDUMP')
-        da_params = {
-            'specie': '1',
-            'time_step': 1,
-            'step_skip': 1,
-            'framework_indices':[]
-        }
+        da_params = {'specie': '1', 'time_step': 1, 'step_skip': 1, 'framework_indices': []}
         data = parser.MDAnalysisParser(xd, **da_params)
         assert_almost_equal(data.time_step, 1)
         assert_almost_equal(data.step_skip, 1)
