### 2023-11-22 #5984 @hmacdope
Kinisi will be extremely useful to those studying transport coefficients in molecular simulation, filling a gap in the existing ecosystem with a well designed and executed package. In particular methods to estimate errors effectively and compatibility with many formats and packages are to be commended.

@arm61 I have one major question and then a few other suggestions.

How are periodic boundaries handled in `kinisi`? This was not obvious from the documentation and is crucial for correctly determining the displacements. If kinisi does not do any unwrapping of periodic trajectories, it needs to be very clear in documentation that the user needs to perform this step themselves prior to feeding into kinisi API.

- I would love to see a conda-forge package, which is very easy to do. I know things can take a little while on `staged-recipies` so even if you just started the process I would be happy.
- You could consider using `statsmodels.stats.correlation_tools.cov_nearest` rather than implementing positive semidefinite covariance matrices yourself, but up to you.
- In the docs, under *Methodology* it is unclear from a first look that a user shouldn't copy the code in the Methodology section rather than using the API. Perhaps if you mapped each bit of methodology to the corresponding classes that would be helpful.

Once the above points are addressed kinisi is thoroughly deserving of publication in JOSS.
