+++ pyproject.toml
@@ -49,7 +49,8 @@ dev = [
     'pytest',
     'pytest-cov',
     'ase',
-    'h5py'
+    'h5py',
+    'pyblock'
 ]
 docs = [
     'pymatgen',
