### 2023-11-23 #5984 @arm61
Thanks for the comments @hmacdope! All great suggestions 😄. I will gradually work through them over the next couple of days. Starting with the `cov_nearest` function, I must admit I was ignorant and not aware of this option. I have tested and if gives the same result as my handmade implementation. Therefore, I have updated the code (https://github.com/bjmorgan/kinisi/pull/48) to use the `statsmodels` function to reduce the overhead.
