### 2023-11-23 #5984 @arm61
I must admit that this is my first time making a condo recipe but hopefully I have got it right. https://github.com/conda-forge/staged-recipes/pull/24572
