### 2023-11-23 #5984 @arm61
Finally, for the methodology comment. The design of `kinisi` API is such that much of the what is discussed in the methodology section is abstracted away from the user. Therefore, there is no real mapping between the methodology and functions/classes that are user facing. Therefore, I have added a comment to the start of the methodology section (https://github.com/bjmorgan/kinisi/commit/cb370ed48de8bcf862de0b4891ee0caf9d04a580) making clear the purpose of the section and pointing the reader to the notebooks to see how to use `kinisi`.

I hope that is acceptable @hmacdope 😅. Thanks again for the great comments.
