+++ docs/source/faq.md
@@ -42,7 +42,7 @@
 - How are trajectories unwrapped?
 
   > When calculating displacements, `kinisi` uses a simple heuristic to unwrap trajectories. 
-  > If the displacement between two steps, is more significant than half the simulation cell length, `kinisi` wraps that
+  > If the displacement between two steps, is greater than half the simulation cell length, `kinisi` wraps that
   > displacement. This scheme assumes that no particle moves more than one cell between steps. Therefore, it requires that
   > enough simulation data is provided to `kinisi`. This process is performed for each dimension of the trajectory,
   > allowing for any orthorhombic (and cubic) cell. However, this heuristic does not support simulation cells that do not have
