### 2023-11-23 ab48715
Extended framework_indices to all parsers
Former-commit-id: 685aa4c3c6267f4b62ca9aff732cd0aae4b5ca58
