### 2023-11-30 17f770f
Remove abbreviations in  paper.md
