### 2023-11-30 6d1c3da
Added progress bar to get_disp
