### 2023-12-05 8818f17
replaced for loop with einsum
A good explaination of how einsum works can be found here: https://stackoverflow.com/questions/26089893/understanding-numpys-einsum
