+++ kinisi/parser.py
@@ -111,8 +111,8 @@ class Parser:
         d_coords = coords[:, 1:] - coords[:, :-1]
         d_coords = d_coords - np.round(d_coords)
         f_disp = np.cumsum(d_coords, axis=1)
-        latt_reshaped = np.array(latt)[1:]
-        disp = np.einsum('ijk,jkl->jik', f_disp, latt_reshaped)
+        latt = np.array(latt)
+        disp = np.einsum('ijk,jkl->jik', f_disp, latt[1:])
         disp = np.transpose(disp, (1, 0, 2))
         return disp
 
