### 2023-12-08 56a1d52
Swaped slow for loop for faster einsum
