+++ kinisi/parser.py
@@ -109,10 +109,9 @@ class Parser:
         d_coords = coords[:, 1:] - coords[:, :-1]
         d_coords = d_coords - np.round(d_coords)
         f_disp = np.cumsum(d_coords, axis=1)
-        c_disp = []
-        for i in f_disp:
-            c_disp.append([np.dot(d, m) for d, m in zip(i, latt[1:])])
-        disp = np.array(c_disp)
+        latt = np.array(latt)
+        disp = np.einsum('ijk,jkl->jik', f_disp, latt[1:])
+        disp = np.transpose(disp, (1, 0, 2))
         return disp
 
     @staticmethod
