### 2023-12-16 #5984 @arm61
Hey @zhubonan. Sorry about the delay. I have found a small bug in kinisi that I would like to deal with before we do a 1.0.0 release. But with the end of term I have been swamped with teaching, it’ll probably be the new year before I have the time to focus on it. I hope that is okay?
