### 2024-01-07 09dded1
Added new tests for ASE and Pymatgen
Former-commit-id: c984d78c090a2617acb634b628ae81367c8662a8
