+++ kinisi/tests/test_parser.py
@@ -90,6 +90,7 @@ class TestParser(unittest.TestCase):
         assert_equal(disp_3d[0].shape[1], 81)
         assert_equal(disp_3d[0].shape[2], 3)
 
+    #Pymatgen tests with VASP XDATCAR files
     def test_pymatgen_init(self):
         xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_XDATCAR.gz'))
         da_params = {'specie': 'Li', 'time_step': 2.0, 'step_skip': 50}
@@ -122,7 +123,51 @@ class TestParser(unittest.TestCase):
         assert_almost_equal(data.time_step, 20.0)
         assert_almost_equal(data.step_skip, 100)
         assert_equal(data.indices, list(range(xd.natoms[0])))
+    
+    def test_pymatgen_init_with_COG(self):
+        xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_center.XDATCAR'))
+        da_params = {'specie': None, 'time_step': 1.0, 'step_skip': 1, 'specie_indices': [[1, 2, 3]]}
+        data = parser.PymatgenParser(xd.structures, **da_params)
+        assert_almost_equal(data.time_step, 1)
+        assert_almost_equal(data.step_skip, 1)
+        assert_equal(data.indices, [0])
+        assert_almost_equal(data.coords_check, [[[0.269634905, 0.262183827, 0.2]]])
 
+    def test_pymatgen_init_with_COM(self):
+        xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_center.XDATCAR'))
+        da_params = {
+            'specie': None,
+            'time_step': 1.0,
+            'step_skip': 1,
+            'specie_indices': [[1, 2, 3]],
+            'masses': [1, 16, 1]
+        }
+        data = parser.PymatgenParser(xd.structures, **da_params)
+        assert_almost_equal(data.time_step, 1)
+        assert_almost_equal(data.step_skip, 1)
+        assert_equal(data.indices, [0])
+        assert_almost_equal(data.coords_check, [[[0.382421597, 0.2087361, 0.2]]])
+
+    def test_pymatgen_init_with_framwork_indices(self):
+        xd = Xdatcar(os.path.join(os.path.dirname(kinisi.__file__), 'tests/inputs/example_drift.XDATCAR'))
+        da_1_params = {'specie': 'H', 'time_step': 1.0, 'step_skip': 1, 'framework_indices': []}
+        data_1 = parser.PymatgenParser(xd.structures, **da_1_params)
+        assert_almost_equal(data_1.time_step, 1)
+        assert_almost_equal(data_1.step_skip, 1)
+        assert_equal(data_1.indices, [0])
+        assert_equal(data_1.drift_indices, [])
+        assert_equal(data_1.dc[0], np.zeros((4, 3)))
+        da_2_params = {'specie': 'H', 'time_step': 1.0, 'step_skip': 1}
+        data_2 = parser.PymatgenParser(xd.structures, **da_2_params)
+        assert_almost_equal(data_2.time_step, 1)
+        assert_almost_equal(data_2.step_skip, 1)
+        assert_equal(data_2.indices, [0])
+        assert_equal(data_2.drift_indices, list(range(1, 9)))
+        print(data_2.dc[0])
+        disp_array = [[0.0, 0.0, 0.0], [0.2, 0.2, 0.2], [0.4, 0.4, 0.4], [1, 1, 1]]
+        assert_almost_equal(data_2.dc[0], disp_array)
+
+    #ASE tests with ASE traj files
     def test_ase_init(self):
[…]
