### 2024-01-07 20f892d
fixed get_indices
