### 2024-01-07 97b1f44
removed duplicate get_molecules functions
Former-commit-id: d47e764d95b1558d3883542cdb602abf9c4da6b6
