### 2024-01-07 9594740
removed duplicate get_framework functions
