+++ kinisi/diffusion.py
@@ -8,7 +8,7 @@ diffusion coefficient from a material.
 # author: Andrew R. McCluskey (arm61)
 
 import warnings
-from typing import List, Union, Any
+from typing import List, Union
 import numpy as np
 from scipy.stats import normaltest, linregress, multivariate_normal
 from scipy.linalg import pinvh
@@ -16,7 +16,6 @@ from scipy.optimize import minimize, curve_fit
 import scipy.constants as const
 import tqdm
 from uravu.distribution import Distribution
-from sklearn.utils import resample
 from emcee import EnsembleSampler
 from statsmodels.stats.correlation_tools import cov_nearest
 
@@ -65,9 +64,7 @@ class Bootstrap:
         self._n = np.array([])
         self._s = np.array([])
         self._v = np.array([])
-        self._n_bootstrap = np.array([])
-        self._s_bootstrap = np.array([])
-        self._v_bootstrap = np.array([])
+        self._hr = np.array([])
         self._n_o = n_o
         self._ngp = np.array([])
         self._sub_sample_dt = sub_sample_dt
@@ -98,9 +95,7 @@ class Bootstrap:
             'n': self._n,
             's': self._s,
             'v': self._v,
-            'n_bootstrap': self._n_bootstrap,
-            's_bootstrap': self._s_bootstrap,
-            'v_bootstrap': self._v_bootstrap,
+            'hr': self._hr,
             'sub_sample_dt': self._sub_sample_dt,
             'dimension': self._dimension,
             'ngp': self._ngp,
@@ -150,9 +145,7 @@ class Bootstrap:
         boot._n = my_dict['n']
         boot._s = my_dict['s']
         boot._v = my_dict['v']
-        boot._n_bootstrap = my_dict['n_bootstrap']
-        boot._s_bootstrap = my_dict['s_bootstrap']
-        boot._v_bootstrap = my_dict['v_bootstrap']
+        boot._hr = my_dict['hr']
         boot._n_o = my_dict['n_o']
         boot._ngp = my_dict['ngp']
         if my_dict['distributions'] is not None:
@@ -253,39 +246,6 @@ class Bootstrap:
         return loop
 
     @staticmethod
-    def sample_until_normal(array: np.ndarray,
-                            n_samples: float,
-                            n_resamples: int,
-                            max_resamples: int,
-                            alpha: float = 1e-3,
-                            random_state: np.random.mtrand.RandomState = None) -> Distribution:
-        """
-        Resample from the distribution until a normal distribution is obtained or a maximum is reached.
-
-        Args:
-        :param array: The array to sample from.
-        :param n_samples: Number of samples.
-        :param r_resamples: Number of resamples to perform initially.
-        :param max_resamples: The maximum number of resamples to perform.
-        :param alpha: Level that p-value should be below in :py:func:`scipy.stats.normaltest` for the distribution
-            to be normal. Optional, default is :py:attr:`1e-3`.
-        :param random_state: A :py:attr:`RandomState` object to be used to ensure reproducibility. Optional,
-            default is :py:attr:`None`.
-
-        :return: The resampled distribution.
-        """
-        # values = _bayesian_bootstrap(array, n_samples, n_resamples, random_state)
-        values = _bootstrap(array, n_samples, n_resamples, random_state)
-        p_value = normaltest(values)[1]
-        while p_value < alpha and len(values) < max_resamples:
-            # values += _bayesian_bootstrap(array, n_samples, 100, random_state)
-            values += _bootstrap(array, n_samples, 100, random_state)
-            p_value = normaltest(values)[1]
-        if len(values) >= max_resamples:
-            warnings.warn("The maximum number of resamples has been reached, and the distribution is not yet normal.")
-        return Distribution(values)
-
-    @staticmethod
[…]
