+++ kinisi/tests/test_diffusion.py
@@ -15,7 +15,7 @@ import warnings
 import numpy as np
 from tqdm import tqdm
 from numpy.testing import assert_almost_equal, assert_equal
-from kinisi.diffusion import Bootstrap, MSDBootstrap, MSTDBootstrap, MSCDBootstrap, _bootstrap
+from kinisi.diffusion import Bootstrap, MSDBootstrap, MSTDBootstrap, MSCDBootstrap
 from uravu.distribution import Distribution
 
 RNG = np.random.RandomState(43)
@@ -36,7 +36,6 @@ class TestBootstrap(unittest.TestCase):
             assert_almost_equal(a._displacements[i], b._displacements[i])
         assert_almost_equal(a._delta_t, b._delta_t)
         assert a._max_obs == b._max_obs
-        assert a._distributions == b._distributions
         assert a.dt.size == b.dt.size
         assert isinstance(b.dt, np.ndarray)
         assert b._diffusion_coefficient is None
@@ -51,7 +50,6 @@ class TestBootstrap(unittest.TestCase):
             assert_almost_equal(bs._displacements[i], d)
         assert_almost_equal(bs._delta_t, dt)
         assert bs._max_obs == 20
-        assert bs._distributions == []
         assert bs.dt.size == 0
         assert isinstance(bs.dt, np.ndarray)
         assert bs._diffusion_coefficient is None
@@ -66,7 +64,6 @@ class TestBootstrap(unittest.TestCase):
             assert_almost_equal(bs._displacements[i], d)
         assert_almost_equal(bs._delta_t, dt[::2])
         assert bs._max_obs == 20
-        assert bs._distributions == []
         assert bs.dt.size == 0
         assert isinstance(bs.dt, np.ndarray)
         assert bs._diffusion_coefficient is None
@@ -80,34 +77,12 @@ class TestBootstrap(unittest.TestCase):
         result = Bootstrap.iterator(False, range(10))
         assert isinstance(result, range)
 
-    def test_sample_until_normal(self):
-        distro1 = Bootstrap.sample_until_normal(RNG.normal(0, 1, size=2000),
-                                                50,
-                                                1000,
-                                                100000,
-                                                random_state=np.random.RandomState(1))
-        assert distro1.size == 1000
-
-    def test_sample_until_normal_random(self):
-        with warnings.catch_warnings(record=True) as _:
-            distro1 = Bootstrap.sample_until_normal(np.arange(1, 10, 1),
-                                                    5,
-                                                    100,
-                                                    100,
-                                                    random_state=np.random.RandomState(0))
-            distro2 = Bootstrap.sample_until_normal(np.arange(1, 10, 1),
-                                                    5,
-                                                    100,
-                                                    100,
-                                                    random_state=np.random.RandomState(0))
-            assert_almost_equal(distro1.samples, distro2.samples)
-
     def test_ngp_calculation(self):
         result = Bootstrap.ngp_calculation(np.array([1, 2, 3]))
         assert_almost_equal(result, -0.3)
 
 
-class TesMSDBootstrap(unittest.TestCase):
+class TestMSDBootstrap(unittest.TestCase):
     """
     Tests for the diffusion.MSDBootstrap class.
     """
@@ -127,8 +102,6 @@ class TesMSDBootstrap(unittest.TestCase):
             assert_equal(a.s, b.s)
             assert_equal(a.v, b.v)
             assert_equal(a.ngp, b.ngp)
-            for i, d in enumerate(a._distributions):
[…]
