+++ kinisi/tests/test_diffusion.py
@@ -15,23 +15,23 @@ import warnings
 import numpy as np
 from tqdm import tqdm
 from numpy.testing import assert_almost_equal, assert_equal
-from kinisi.diffusion import Bootstrap, MSDBootstrap, MSTDBootstrap, MSCDBootstrap
+from kinisi.diffusion import Diffusion, MSDDiffusion, MSTDDiffusion, MSCDDiffusion
 from uravu.distribution import Distribution
 
 RNG = np.random.RandomState(43)
 
 
-class TestBootstrap(unittest.TestCase):
+class TestDiffusion(unittest.TestCase):
     """
-    Tests for the diffusion.Bootstrap class.
+    Tests for the diffusion.Diffusion class.
     """
 
     def test_dictionary_roundtrip(self):
         disp_3d = [RNG.randn(100, i, 3) for i in range(20, 10, -1)]
         n_o = np.ones(len(disp_3d)) * 100
         dt = np.linspace(100, 1000, 10)
-        a = Bootstrap(dt, disp_3d, n_o)
-        b = Bootstrap.from_dict(a.to_dict())
+        a = Diffusion(dt, disp_3d, n_o)
+        b = Diffusion.from_dict(a.to_dict())
         for i, d in enumerate(disp_3d):
             assert_almost_equal(a._displacements[i], b._displacements[i])
         assert_almost_equal(a._delta_t, b._delta_t)
@@ -45,7 +45,7 @@ class TestBootstrap(unittest.TestCase):
         disp_3d = [RNG.randn(100, i, 3) for i in range(20, 10, -1)]
         n_o = np.ones(len(disp_3d)) * 100
         dt = np.linspace(100, 1000, 10)
-        bs = Bootstrap(dt, disp_3d, n_o)
+        bs = Diffusion(dt, disp_3d, n_o)
         for i, d in enumerate(disp_3d):
             assert_almost_equal(bs._displacements[i], d)
         assert_almost_equal(bs._delta_t, dt)
@@ -59,7 +59,7 @@ class TestBootstrap(unittest.TestCase):
         disp_3d = [RNG.randn(100, i, 3) for i in range(20, 10, -1)]
         n_o = np.ones(len(disp_3d)) * 100
         dt = np.linspace(100, 1000, 10)
-        bs = Bootstrap(dt, disp_3d, n_o, sub_sample_dt=2)
+        bs = Diffusion(dt, disp_3d, n_o, sub_sample_dt=2)
         for i, d in enumerate(disp_3d[::2]):
             assert_almost_equal(bs._displacements[i], d)
         assert_almost_equal(bs._delta_t, dt[::2])
@@ -70,21 +70,21 @@ class TestBootstrap(unittest.TestCase):
         assert bs.intercept is None
 
     def test_iterator_true(self):
-        result = Bootstrap.iterator(True, range(10))
+        result = Diffusion.iterator(True, range(10))
         assert isinstance(result, tqdm)
 
     def test_iterator_false(self):
-        result = Bootstrap.iterator(False, range(10))
+        result = Diffusion.iterator(False, range(10))
         assert isinstance(result, range)
 
     def test_ngp_calculation(self):
-        result = Bootstrap.ngp_calculation(np.array([1, 2, 3]))
+        result = Diffusion.ngp_calculation(np.array([1, 2, 3]))
         assert_almost_equal(result, -0.3)
 
 
-class TestMSDBootstrap(unittest.TestCase):
+class TestMSDDiffusion(unittest.TestCase):
     """
-    Tests for the diffusion.MSDBootstrap class.
+    Tests for the diffusion.MSDDiffusion class.
     """
 
     def test_dictionary_roundtrip(self):
@@ -92,8 +92,8 @@ class TestMSDBootstrap(unittest.TestCase):
[…]
