+++ kinisi/jump_diffusion_analyzer.py
@@ -17,7 +17,7 @@ class JumpDiffusionAnalyzer(Analyzer):
     """
     The :py:class:`kinisi.analyze.JumpDiffusionAnalyzer` class performs analysis of collective diffusion
     relationships in materials.
-    This is achieved through the application of a bootstrapping methodology to obtain the most statistically
+    This is achieved through the application of a Bayesian regression methodology to obtain the most statistically
     accurate values for total mean squared displacement uncertainty and covariance.
     The time-dependence of the MSTD is then modelled in a generalised least squares fashion to obtain the jump
     diffusion coefficient and offset using Markov chain Monte Carlo maximum likelihood sampling.
@@ -27,8 +27,8 @@ class JumpDiffusionAnalyzer(Analyzer):
         There is one array in the list for each delta_t value. Note: it is necessary to use a list of arrays as
         the number of observations is not necessary the same at each data point.
     :param volume: The volume of the simulation cell.
-    :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.DiffBootstrap` object. See
-        the appropriate documentation for more guidance on this. Optional, default is the default bootstrap parameters.
+    :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.DiffDiffusion` object. See
+        the appropriate documentation for more guidance on this. Optional, default is the default diffusion parameters.
     """
 
     def __init__(self, delta_t: np.ndarray, disp_3d: List[np.ndarray], n_o: np.ndarray, volume: float):
@@ -53,7 +53,7 @@ class JumpDiffusionAnalyzer(Analyzer):
         :return: New :py:class:`DiffusionAnalyzer` object.
         """
         jdiff_anal = cls(my_dict['delta_t'], my_dict['disp_3d'], my_dict['n_o'], my_dict['volume'])
-        jdiff_anal._diff = diffusion.Bootstrap.from_dict(my_dict['diff'])
+        jdiff_anal._diff = diffusion.Diffusion.from_dict(my_dict['diff'])
         return jdiff_anal
 
     @classmethod
@@ -75,15 +75,15 @@ class JumpDiffusionAnalyzer(Analyzer):
             these constitute a series of :py:attr:`consecutive` trajectories or a series of :py:attr:`identical`
             starting points with different random seeds, in which case the `dtype` should be either
             :py:attr:`consecutive` or :py:attr:`identical`.
-        :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.MSDBootstrap` object. See the
-            appropriate documentation for more guidance on this. Optional, default is the default bootstrap parameters.
+        :param uncertainty_params: The parameters for the :py:class:`kinisi.diffusion.MSDDiffusion` object. See the
+            appropriate documentation for more guidance on this. Optional, default is the default diffusion parameters.
 
         :return: Relevant :py:class:`JumpDiffusionAnalyzer` object.
         """
         if uncertainty_params is None:
             uncertainty_params = {}
         jdiff_anal = super()._from_pymatgen(trajectory, parser_params, dtype=dtype)
-        jdiff_anal._diff = diffusion.MSTDBootstrap(jdiff_anal._delta_t, jdiff_anal._disp_3d, jdiff_anal._n_o,
+        jdiff_anal._diff = diffusion.MSTDDiffusion(jdiff_anal._delta_t, jdiff_anal._disp_3d, jdiff_anal._n_o,
                                                    **uncertainty_params)
         return jdiff_anal
 
@@ -105,15 +105,15 @@ class JumpDiffusionAnalyzer(Analyzer):
             these constitute a series of :py:attr:`consecutive` trajectories or a series of :py:attr:`identical`
[…]

[…]
