### 2024-01-25 #5984 @arm61
Hey @zhubonan, sorry again for the delay. The final thing we wanted to fix has an open PR (https://github.com/bjmorgan/kinisi/pull/58). Hopefully, it will be reviewed next week, and once merged, I will sort the rest of the bits and bobs out.
