### 2024-01-28 #5984 @zhubonan
OK, looking forwarding to see it fixed 🚀
BTW, I actually think it is OK to release and save the archive without the fix, if it takes too long, since the reviews process were done before the fix is applied.
You can always make a new release of the code and link the new DOI with the Zenodo record.
