### 2024-02-09 1d49248
Make normalisation optional
