### 2024-02-11 32c5523
Make magnitude
