+++ paper/paper.md
@@ -33,7 +33,7 @@ affiliations:
    index: 4
  - name: The Faraday Institution, Quad One, Harwell Science and Innovation Campus, Didcot, OX11 0RA, United Kingdom
    index: 5
-date: 16 August 2023
+date: 12 February 2024
 bibliography: paper.bib
 ---
 
