+++ paper/paper.md
@@ -68,5 +68,5 @@ A list of publications where `kinisi` has been used in the analysis of simulatio
 # Acknowledgements
 
 The authors thank all of the users of `kinisi` for contributing feedback, suggesting new features and filing bug reports. 
-S.W.C., A.G.S. and B.J.M. acknowledge the support of the Faraday Institution (grant numbers FIRG016, FIRG017).
+S.W.C., A.G.S., and B.J.M. acknowledge the support of the Faraday Institution (grant numbers FIRG016 and FIRG017).
 B.J.M. acknowledges support from the Royal Society (UF130329 and URF\textbackslash R\textbackslash 191006). 
