### 2024-02-12 b2ae3d4
Update CITATION.cff
