+++ CITATION.cff
@@ -1,15 +1,23 @@
 cff-version: 1.2.0
 abstract: "Uncertainty analysis and model comparison for atomistic molecular dynamics."
 authors: 
-- affiliation: "European Spallation Source ERIC"
+- affiliation: "Centre for Computational Chemistry, School of Chemistry, University of Bristol and European Spallation Source ERIC"
   family-names: McCluskey
   given-names: "Andrew R."
   orcid: "https://orcid.org/0000-0003-3381-5911"
-- affilation: "Department of Chemistry, University College London"
+- affilation: "School of Chemistry, University of Birmingham"
   family-names: Squires
   given-names: "Alex G."
   orcid: "https://orcid.org/0000-0001-6967-3690"
-- affiliation: "Department of Chemistry, University of Bath"
+- affilation: "Centre for Computational Chemistry, School of Chemistry, University of Bristol"
+  family-names: Dunn
+  given-names: "Josh"
+  orcid: "https://orcid.org/0000-0003-2659-0806"
+- affilation: "Department of Chemistry, University of Bath and The Faraday Institution"
+  family-names: Coles
+  given-names: "Samuel W."
+  orcid: "https://orcid.org/0000-0001-9722-5676"
+- affiliation: "Department of Chemistry, University of Bath and The Faraday Institution"
   family-names: Morgan
   given-names: "Benjamin J."
   orcid: "https://orcid.org/0000-0002-3056-8233"
@@ -17,8 +25,8 @@ date-released: 2022-07-18
 license: MIT
 message: "If you use this software, please cite it using these metadata."
 repository-code: "https://github.com/bjmorgan/kinisi"
-title: kinisi
-version: "0.7.0"
+title: "kinisi: Bayesian analysis of mass transport from molecular dynamics simulations"
+version: "1.0.0"
 preferred-citation:
   type: article
   title: "Accurate Estimation of Diffusion Coefficients and their Uncertainties from Computer Simulation"
