### 2024-02-12 2c3b55e
Update CITATION.cff
