+++ CITATION.cff
@@ -21,7 +21,7 @@ authors:
   family-names: Morgan
   given-names: "Benjamin J."
   orcid: "https://orcid.org/0000-0002-3056-8233"
-date-released: 2022-07-18
+date-released: 2024-02-12
 license: MIT
 message: "If you use this software, please cite it using these metadata."
 repository-code: "https://github.com/bjmorgan/kinisi"
