+++ paper/paper.bib
@@ -35,7 +35,7 @@
   volume = {4},
   number = {43},
   pages = {1864},
-  author = {{Foreman-Mackey}, D and {Farr}, W. M. and {Sinha}, M. and {Archibald}, A. M. and {Hogg}, D. W. and {Sanders}, J. S. and {Zuntz}, J. and {Williams}, P. K. G. and {Nelson}, A. R. J. and {de Val-Borro}, M. and {Erhardt}, T. and {Pashchenko}, I. and {Pla}, O. A.},
+  author = {{Foreman-Mackey}, D. and {Farr}, W. M. and {Sinha}, M. and {Archibald}, A. M. and {Hogg}, D. W. and {Sanders}, J. S. and {Zuntz}, J. and {Williams}, P. K. G. and {Nelson}, A. R. J. and {de Val-Borro}, M. and {Erhardt}, T. and {Pashchenko}, I. and {Pla}, O. A.},
   title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC},
   journal = {J. Open Source Softw.}
 }
@@ -129,7 +129,7 @@
   volume = {5},
   number = {1},
   pages = {65--80},
-  author = {Jonathan Goodman and Jonathan Weare},
+  author = {Goodman, J. and Weare, J.},
   title = {Ensemble samplers with affine invariance},
   journal = {Communications in Applied Mathematics and Computational Science}
 }
@@ -148,7 +148,7 @@
   year = {2014},
   volume = {112},
   number = {14},
-  author = {Benjamin J. Morgan and Paul A. Madden},
+  author = {B. J. Morgan and P. A. Madden},
   title = {Relationships between Atomic Diffusion Mechanisms and Ensemble Transport Coefficients in Crystalline Polymorphs},
   journal = {Phys. Rev. Lett.}
 }
