### 2024-02-13 #5984 @arm61
Annoyingly, the Zenodo archive didn't read the affiliations correctly from the CITATION.cff file. I will ask @bjmorgan to sort the Zenodo deposit as he is the owner.
