+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
     affiliation: "1,2"
   - name: Alexander G. Squires
     orcid: 0000-0001-6967-3690
-    affiliation: "3"
+    affiliation: "3,5"
   - name: Josh Dunn
     orcid: 0000-0003-2659-0806
     affiliation: "1"
