### 2024-02-19 #5984 @Kevin-Mattheus-Moerman
@arm61 as AEiC I will now help to process final steps for acceptance in JOSS. I have checked this review, your repository, the paper, and the archive link. All seems in order so I will now process this for acceptance. I do however have the following recommendation, if you do not already, please consider linking to the CONTRIBUTING.md file from the README and documentation. Currently I can see the file, but from having a quick look I cannot see how folks would end up there other than looking for an CONTRIBUTING.md document. A short section in your README that links to it may be nice.
